+++ src/schemathesis/cli/__init__.py
@@ -9,7 +9,7 @@ from collections import defaultdict
 from dataclasses import dataclass
 from enum import Enum
 from queue import Queue
-from typing import TYPE_CHECKING, Any, Callable, Generator, Iterable, Literal, NoReturn, Sequence, Type, cast
+from typing import TYPE_CHECKING, Any, Callable, Generator, Iterable, Literal, NoReturn, Sequence, cast
 from urllib.parse import urlparse
 
 import click
@@ -27,6 +27,7 @@ from ..constants import (
     DEFAULT_STATEFUL_RECURSION_LIMIT,
     EXTENSIONS_DOCUMENTATION_URL,
     HOOKS_MODULE_ENV_VAR,
+    HTTP_METHODS,
     HYPOTHESIS_IN_MEMORY_DATABASE_IDENTIFIER,
     ISSUE_TRACKER_URL,
     WAIT_FOR_SCHEMA_ENV_VAR,
@@ -322,6 +323,16 @@ REPORT_TO_SERVICE = ReportToService()
     callback=callbacks.convert_experimental,
     multiple=True,
     metavar="",
+)
+@grouped_option(
+    "--experimental-coverage-unexpected-methods",
+    "coverage_unexpected_methods",
+    help="HTTP methods to use when generating test cases with methods not specified in the API during the coverage phase.",
+    type=CsvChoice(sorted(HTTP_METHODS), case_sensitive=False),
+    callback=callbacks.convert_http_methods,
+    metavar="",
+    default=None,
+    envvar="SCHEMATHESIS_EXPERIMENTAL_COVERAGE_UNEXPECTED_METHODS",
 )
 @grouped_option(
     "--experimental-no-failfast",
@@ -874,6 +885,7 @@ def run(
     set_path: dict[str, str],
     experiments: list,
     no_failfast: bool,
+    coverage_unexpected_methods: set[str] | None,
     missing_required_header_allowed_statuses: list[str],
     positive_data_acceptance_allowed_statuses: list[str],
     negative_data_rejection_allowed_statuses: list[str],
@@ -1000,6 +1012,7 @@ def run(
         graphql_allow_null=generation_graphql_allow_null,
         codec=generation_codec,
         with_security_parameters=generation_with_security_parameters,
+        unexpected_methods=coverage_unexpected_methods,
     )
 
     report: ReportToService | click.utils.LazyFile | None