+++ src/schemathesis/cli/callbacks.py
@@ -344,6 +344,14 @@ def convert_checks(ctx: click.core.Context, param: click.core.Parameter, value: 
     return reduce(operator.iadd, value, [])
 
 
+def convert_http_methods(
+    ctx: click.core.Context, param: click.core.Parameter, value: list[str] | None
+) -> set[str] | None:
+    if value is None:
+        return value
+    return {item.lower() for item in value}
+
+
 def convert_status_codes(
     ctx: click.core.Context, param: click.core.Parameter, value: list[str] | None
 ) -> list[str] | None: