+++ docs/changelog.rst
@@ -4,6 +4,10 @@ Changelog
 :version:`Unreleased <v3.39.11...HEAD>` - TBD
 ---------------------------------------------
 
+**Added**
+
+- `--experimental-coverage-unspecified-methods` CLI option that accept a comma separated list of HTTP methods to use when generating test cases with methods not specified in the API during the **coverage** phase.
+
 .. _v3.39.11:
 
 :version:`3.39.11 <v3.39.10...v3.39.11>` - 2025-02-23